# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "### FIAP Tech Challenge Phase 3 - 10SOAT - G22",
        "title": "Fast Food API v3",
        "contact": {},
        "version": "1"
    },
    "basePath": "/api/v1",
    "paths": {
...
        "/orders": {
            "get": {
                "description": "List all orders\n## Order list is sorted by:\n- **Status** in **descending** order (`READY` \u003e `PREPARING` \u003e `RECEIVED` \u003e `PENDING` \u003e `OPEN`)\n- **Created date** (CreatedAt) in **ascending** order (oldest first)\nObs: Status CANCELLED and COMPLETED are not included in the list by default",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "orders"
                ],
                "summary": "List orders (Reference TC-1 2.b.vi; TC-2 1.a.iv)",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Filter by customer ID",
                        "name": "customer_id",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Filter by status (Accept many), options: \u003csub\u003eOPEN, PENDING, RECEIVED, PREPARING, READY\u003c/sub\u003e, ex: \u003csub\u003ePENDING\u003c/sub\u003e or \u003csub\u003eOPEN,PENDING\u003c/sub\u003e",
                        "name": "status",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "default": "CANCELLED,COMPLETED",
                        "description": "Exclude by status (Accept many), options: \u003csub\u003eNONE, OPEN, PENDING, RECEIVED, PREPARING, READY, CANCELLED, COMPLETED\u003c/sub\u003e, ex: \u003csub\u003eCANCELLED\u003c/sub\u003e or \u003csub\u003eCANCELLED,COMPLETED\u003c/sub\u003e (default)",
                        "name": "status_exclude",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "default": "status:d,created_at",
                        "description": "Sort by field (Accept many). Use `\u003cfield_name\u003e:d` for descending, and the default order is ascending",
                        "name": "sort",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "default": 1,
                        "description": "Page number",
                        "name": "page",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "default": 10,
                        "description": "Items per page",
                        "name": "limit",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/presenter.OrderJsonPaginatedResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/middleware.ErrorJsonResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/middleware.ErrorJsonResponse"
                        }
                    }
                }
            },
            "post": {
                "description": "Creates a new order",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "orders"
                ],
                "summary": "Create order",
                "parameters": [
                    {
                        "description": "Order data",
                        "name": "order",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/request.CreateOrderBodyRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/presenter.OrderJsonResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/middleware.ErrorJsonResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/middleware.ErrorJsonResponse"
                        }
                    }
                }
            }
        },
        "/orders/histories": {
            "get": {
                "description": "List all order histories",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "orders"
                ],
                "summary": "List order histories",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Filter by order_id",
                        "name": "order_id",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Filter by status. Available options: OPEN, CANCELLED, PENDING, RECEIVED, PREPARING, READY, COMPLETED",
                        "name": "status",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "default": 1,
                        "description": "Page number",
                        "name": "page",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "default": 10,
                        "description": "Items per page",
                        "name": "limit",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/presenter.OrderHistoryJsonPaginatedResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/middleware.ErrorJsonResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/middleware.ErrorJsonResponse"
                        }
                    }
                }
            }
        },
        "/orders/histories/{id}": {
            "get": {
                "description": "Search for a order history by ID",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "orders"
                ],
                "summary": "Get order history",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "OrderHistory ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/presenter.OrderHistoryJsonResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/middleware.ErrorJsonResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/middleware.ErrorJsonResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/middleware.ErrorJsonResponse"
                        }
                    }
                }
            },
            "delete": {
                "description": "Deletes a order history by ID",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "orders"
                ],
                "summary": "Delete order history",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "OrderHistory ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/presenter.OrderHistoryJsonResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/middleware.ErrorJsonResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/middleware.ErrorJsonResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/middleware.ErrorJsonResponse"
                        }
                    }
                }
            }
        },
...
        "/orders/{id}": {
            "get": {
                "description": "Search for a order by ID",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "orders"
                ],
                "summary": "Get order",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Order ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/presenter.OrderJsonResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/middleware.ErrorJsonResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/middleware.ErrorJsonResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/middleware.ErrorJsonResponse"
                        }
                    }
                }
            },
            "put": {
                "description": "Update an existing order\nThe status are: **OPEN**, **CANCELLED**, **PENDING**, **RECEIVED**, **PREPARING**, **READY**, **COMPLETED**\n## Transition of status:\n- OPEN      -\u003e CANCELLED || PENDING\n- CANCELLED -\u003e {},\n- PENDING   -\u003e OPEN || RECEIVED\n- RECEIVED  -\u003e PREPARING\n- PREPARING -\u003e READY\n- READY     -\u003e COMPLETED\n- COMPLETED -\u003e {}",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "orders"
                ],
                "summary": "Update order",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Order ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Order data",
                        "name": "order",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/request.UpdateOrderBodyRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/presenter.OrderJsonResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/middleware.ErrorJsonResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/middleware.ErrorJsonResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/middleware.ErrorJsonResponse"
                        }
                    }
                }
            },
            "delete": {
                "description": "Deletes a order by ID",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "orders"
                ],
                "summary": "Delete order",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Order ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/presenter.OrderJsonResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/middleware.ErrorJsonResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/middleware.ErrorJsonResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/middleware.ErrorJsonResponse"
                        }
                    }
                }
            },
            "patch": {
                "description": "Partially updates an existing order\nThe status are: **OPEN**, **CANCELLED**, **PENDING**, **RECEIVED**, **PREPARING**, **READY**, **COMPLETED**\n## Transition of status:\n- OPEN      -\u003e CANCELLED || PENDING\n- CANCELLED -\u003e {},\n- PENDING   -\u003e OPEN || RECEIVED\n- RECEIVED  -\u003e PREPARING\n- PREPARING -\u003e READY\n- READY     -\u003e COMPLETED\n- COMPLETED -\u003e {}",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "orders"
                ],
                "summary": "Partial update order (Reference TC-2 1.a.v)",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Order ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Order data",
                        "name": "order",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/request.UpdateOrderPartilRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/presenter.OrderJsonResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/middleware.ErrorJsonResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/middleware.ErrorJsonResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/middleware.ErrorJsonResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "middleware.ErrorJsonResponse": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer",
                    "example": 400
                },
                "message": {
                    "type": "string",
                    "example": "Bad Request"
                }
            }
        },
...
        "presenter.CustomerJsonResponse": {
            "type": "object",
            "properties": {
                "cpf": {
                    "type": "string",
                    "example": "123.456.789-00"
                },
                "created_at": {
                    "type": "string",
                    "example": "2024-02-09T10:00:00Z"
                },
                "email": {
                    "type": "string",
                    "example": "john.doe@email.com"
                },
                "id": {
                    "type": "integer",
                    "example": 1
                },
                "name": {
                    "type": "string",
                    "example": "John Doe"
                },
                "updated_at": {
                    "type": "string",
                    "example": "2024-02-09T10:00:00Z"
                }
            }
        },
        "presenter.OrderHistoryJsonPaginatedResponse": {
            "type": "object",
            "properties": {
                "limit": {
                    "type": "integer",
                    "example": 10
                },
                "order_histories": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/presenter.OrderHistoryJsonResponse"
                    }
                },
                "page": {
                    "type": "integer",
                    "example": 1
                },
                "total": {
                    "type": "integer",
                    "example": 100
                }
            }
        },
        "presenter.OrderHistoryJsonResponse": {
            "type": "object",
            "properties": {
                "created_at": {
                    "type": "string",
                    "example": "2024-02-09T10:00:00Z"
                },
                "id": {
                    "type": "integer",
                    "example": 1
                },
                "order_id": {
                    "type": "integer",
                    "example": 1
                },
                "staff_id": {
                    "type": "integer",
                    "example": 1
                },
                "status": {
                    "type": "string",
                    "example": "OPEN, CANCELLED, PENDING, RECEIVED, PREPARING, READY, COMPLETED"
                }
            }
        },
        "presenter.OrderJsonPaginatedResponse": {
            "type": "object",
            "properties": {
                "limit": {
                    "type": "integer",
                    "example": 10
                },
                "orders": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/presenter.OrderJsonResponse"
                    }
                },
                "page": {
                    "type": "integer",
                    "example": 1
                },
                "total": {
                    "type": "integer",
                    "example": 100
                }
            }
        },
        "presenter.OrderJsonResponse": {
            "type": "object",
            "properties": {
                "created_at": {
                    "type": "string",
                    "example": "2024-02-09T10:00:00Z"
                },
                "customer": {
                    "$ref": "#/definitions/presenter.CustomerJsonResponse"
                },
                "customer_id": {
                    "type": "integer",
                    "example": 1
                },
                "id": {
                    "type": "integer"
                },
                "products": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/presenter.ProductsJsonResponse"
                    }
                },
                "status": {
                    "type": "string",
                    "example": "PENDING"
                },
                "total_bill": {
                    "type": "string",
                    "example": "100.00"
                },
                "updated_at": {
                    "type": "string",
                    "example": "2024-02-09T10:00:00Z"
                }
            }
        },
...
        "presenter.ProductsJsonResponse": {
            "type": "object",
            "properties": {
                "category_id": {
                    "type": "integer",
                    "example": 1
                },
                "created_at": {
                    "type": "string",
                    "example": "2024-02-09T10:00:00Z"
                },
                "description": {
                    "type": "string",
                    "example": "Description of product A"
                },
                "id": {
                    "type": "integer",
                    "example": 1
                },
                "name": {
                    "type": "string",
                    "example": "Product A"
                },
                "price": {
                    "type": "number",
                    "example": 99.99
                },
                "quantity": {
                    "type": "integer"
                },
                "updated_at": {
                    "type": "string",
                    "example": "2024-02-09T10:00:00Z"
                }
            }
        },
...
        "request.CreateOrderBodyRequest": {
            "type": "object",
            "required": [
                "customer_id"
            ],
            "properties": {
                "customer_id": {
                    "type": "integer",
                    "example": 1
                }
            }
        },
...
        "request.UpdateOrderBodyRequest": {
            "type": "object",
            "required": [
                "customer_id",
                "status"
            ],
            "properties": {
                "customer_id": {
                    "type": "integer",
                    "example": 1
                },
                "staff_id": {
                    "description": "StaffID is only required when status is PREPARING, READY or COMPLETED",
                    "type": "integer",
                    "example": 1
                },
                "status": {
                    "allOf": [
                        {
                            "$ref": "#/definitions/valueobject.OrderStatus"
                        }
                    ],
                    "example": "PENDING"
                }
            }
        },
        "request.UpdateOrderPartilRequest": {
            "type": "object",
            "properties": {
                "staff_id": {
                    "description": "StaffID is only required when status is PREPARING, READY or COMPLETED",
                    "type": "integer",
                    "example": 1
                },
                "status": {
                    "allOf": [
                        {
                            "$ref": "#/definitions/valueobject.OrderStatus"
                        }
                    ],
                    "example": "PENDING"
                }
            }
        },
...
        "valueobject.OrderStatus": {
            "type": "string",
            "enum": [
                "OPEN",
                "CANCELLED",
                "PENDING",
                "RECEIVED",
                "PREPARING",
                "READY",
                "COMPLETED",
                "UNDEFINDED"
            ],
            "x-enum-varnames": [
                "OPEN",
                "CANCELLED",
                "PENDING",
                "RECEIVED",
                "PREPARING",
                "READY",
                "COMPLETED",
                "UNDEFINDED"
            ]
        },
...
    },
    "securityDefinitions": {
        "BearerAuth": {
            "description": "Type \"Bearer\" followed by a space and the access token.",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    },
...
}